requirements.txt
tornado==4.2.1
PyMySQL==0.7.9
SQLAlchemy==1.1.4
redis==2.10.3
oss2==2.2.0
pillow==3.4.1
mutagen==1.34.1
pyqrcode==1.2.1
